Fixed saving icons with newest PyObjC
Ignore-this: b143d6813dba13ffb3dc08dc1b75b466

When extracting the data out of an image file, we need to get the actual bytes to write to disk,
this is done with the .bytes() call. On PyObjC >= 2.3 .bytes() returns a memoryview object instead
of the actual bytes, so .tobytes() needs to be called in order to fetch the actual image bytes.

darcs-hash:20130603115023-56b4d-f28ba5dae21beed6a1b0230b28c8bccf78b47e68